Move finalty tracker inherents into primitives
