+++ client/finality-grandpa/Cargo.toml
@@ -25,7 +25,7 @@ client = { package = "substrate-client", path = "../" }
 header-metadata = { package = "substrate-header-metadata", path = "../header-metadata" }
 inherents = { package = "substrate-inherents", path = "../../primitives/inherents" }
 network = { package = "substrate-network", path = "../network" }
-paint-finality-tracker = { path = "../../paint/finality-tracker" }
+sp-finality-tracker = { path = "../../primitives/finality-tracker" }
 fg_primitives = { package = "substrate-finality-grandpa-primitives", path = "../../primitives/finality-grandpa" }
 grandpa = { package = "finality-grandpa", version = "0.9.0", features = ["derive-codec"] }
