+++ paint/finality-tracker/Cargo.toml
@@ -10,6 +10,7 @@ codec = { package = "parity-scale-codec", version = "1.0.0", default-features = 
 inherents = { package = "substrate-inherents", path = "../../primitives/inherents", default-features = false }
 rstd = { package = "sr-std", path = "../../primitives/sr-std", default-features = false }
 sr-primitives = { path = "../../primitives/sr-primitives", default-features = false }
+sp-finality-tracker = { path = "../../primitives/finality-tracker", default-features = false }
 support = { package = "paint-support", path = "../support", default-features = false }
 paint-system = { path = "../system", default-features = false }
 impl-trait-for-tuples = "0.1.3"
@@ -27,5 +28,6 @@ std = [
 	"support/std",
 	"sr-primitives/std",
 	"paint-system/std",
+	"sp-finality-tracker/std",
 	"inherents/std",
 ]
