+++ primitives/finality-tracker/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "sp-finality-tracker"
+version = "2.0.0"
+authors = ["Parity Technologies <admin@parity.io>"]
+edition = "2018"
+
+[dependencies]
+codec = { package = "parity-scale-codec", version = "1.0.0", default-features = false }
+inherents = { package = "substrate-inherents", path = "../../primitives/inherents", default-features = false }
+rstd = { package = "sr-std", path = "../../primitives/sr-std", default-features = false }
+
+[features]
+default = ["std"]
+std = [
+	"codec/std",
+	"rstd/std",
+	"inherents/std",
+]
